requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=2.2.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
